rename some type of crop_primitive
