+++ backend/crop_primitive/src/lib.rs
@@ -3,7 +3,7 @@ extern crate serde;
 #[macro_use]
 extern crate derive_more;
 
-pub mod string;
-pub mod vec;
-pub use self::string::NonEmptyString;
-pub use self::vec::NonEmptyVec;
+pub mod non_empty_string;
+pub mod non_empty_vec;
+pub use self::non_empty_string::NonEmptyString;
+pub use self::non_empty_vec::NonEmptyVec;
