+++ backend/crop_primitive/src/non_empty_string.rs
